USD - Fixed slow reload layers
Because of prim collapsing. we could receive multiple resyncs for different prims that were all collapsed into the same prim, resulting in us reloading the asset multiple times. We could also receive an update on the root which was triggering a reload of all the assets when we should have just been updating that particular prim.

#jira none
#rb Daniel.Coelho

#ROBOMERGE-SOURCE: CL 12066198 in //UE4/Release-4.25/... via CL 12066211
#ROBOMERGE-BOT: RELEASE (Release-4.25Plus -> Main) (v657-12064184)

[CL 12078250 by jeanmichel dignard in Main branch]