Fix for ensure messages disappearing in editor log
Calls to PanicFlushLog should only happen on crashes, and not ensures. On ensures on any thread other than the main thread the editor log, which is not thread safe, will loose these messages. On crashes we don't care about the non thread-safe output devices since we will not recover, we are only interested in the log file.

#jira UE-82581
#rb rob.manuszewski
#ushell-cherrypick of 12147436 by Johan.Berg

#ROBOMERGE-OWNER: ryan.gerleve
#ROBOMERGE-AUTHOR: johan.berg
#ROBOMERGE-SOURCE: CL 12192433 in //UE4/Release-4.25/... via CL 12192441 via CL 12192448
#ROBOMERGE-BOT: NETWORKING (Main -> Dev-Networking) (v663-12210973)

[CL 12223023 by johan berg in Dev-Networking branch]