Fix for incorrect callstack in crash reports
When a ensure is reported before a crash the callstack from the crash was added to the ensure stack frames. This was because the debug helper instance used to analyze the minidump is a singleton and not designed to be reused. Use GetNew method to create a new instance of the helper, and delete the old one.

#jira UE-90699
#rb stefan.boberg

[CL 12245123 by Johan Berg in 4.25 branch]