Fix for incorrect callstack in crash reports
When a ensure is reported before a crash the callstack from the crash was added to the ensure stack frames. This was because the debug helper instance used to analyze the minidump is a singleton and not designed to be reused. Use GetNew method to create a new instance of the helper, and delete the old one.

#jira UE-90699
#rb stefan.boberg

#ROBOMERGE-SOURCE: CL 12245123 in //UE4/Release-4.25/... via CL 12245127
#ROBOMERGE-BOT: RELEASE (Release-4.25Plus -> Main) (v667-12241502)

[CL 12245131 by johan berg in Main branch]