Undo changelist 11579245 in //UE4/Release-4.25 as it caused a new bug
"When looking up a blueprint property's name from a guid we now prefer names in the super class, when looking up a guid from a name we still prefer guids in the derived class. No longer copying all guids into derived types on compile. This fixes loss of data in derived types when renaming a property"

#rb dan.oconnor
#jira UE-90825
#rnx

[CL 12249251 by ben hoffman in 4.25 branch]