Fix for incorrect callstack in crash reports
When a ensure is reported before a crash the callstack from the crash was added to the ensure stack frames. This was because the debug helper instance used to analyze the minidump is a singleton and not designed to be reused. Use GetNew method to create a new instance of the helper, and delete the old one.

#jira UE-90699
#rb stefan.boberg

#ROBOMERGE-OWNER: ryan.gerleve
#ROBOMERGE-AUTHOR: johan.berg
#ROBOMERGE-SOURCE: CL 12245123 in //UE4/Release-4.25/... via CL 12245127 via CL 12245131
#ROBOMERGE-BOT: NETWORKING (Main -> Dev-Networking) (v668-12245121)

[CL 12281473 by johan berg in Dev-Networking branch]