The GPU version of the GetNumParticles function in the direct read DI uses the value from the GPU-side instance count buffer, instead of the value which the batcher has.
The value known to the batcher can be one or more frames behind the actual number. This approach returns the accurate value.

Also cleaned up the UnsetShaderParams calls from NiagaraDataSet. It's no longer necessary to unset the buffers when iterated stages are used, since the translator was modified previously to no longer emit code which writes to those buffers for iterated stages.

#rb none
#jira none
#rnx

[CL 12291290 by mihnea balta in 4.25 branch]