Added new External Package feature.
This allows objects to be assigned a package to be saved in that is different from their outer chain. This feature is only available in the Editor.
Assigning an external package to an object will return that package when GetOutermost or the new GetPackage function is called on that object or any objects outered to it.
This assigned package is also considered when calling `IsIn` which now dispatch to IsInObject or IsInPackage according to the parameter's class. IsInObject or IsInPackage should be preferred in the future.
The assigned package is also properly considered when calling GetObjectsWithOuter or ForEachObjectWithOuter. New variant called GetObjectsWithPackage and ForEachObjectWithPackage should be preferred when gathering package content.

This allows to split objects into different packages without affecting the outer chain hierarchy and will support the one file per actor initiative.

#rb Matt.Peters, Robert.Manuszewski

[CL 12371382 by Francis Hurteau in Dev-Core branch]