Fixed crash when using a loop with more than 32 iterations in custom HLSL.
The default threshold for loop unrolling was 32, so the compiler tried to make a dynamic loop when we had more iterations. The vector VM doesn't support that, so the codegen was never designed to understand dynamic loops, leading to an assertion failure. Increased the threshold to a large value, so we never hit this problem in practice.

#rnx
#jira none
#rb none

[CL 12422302 by mihnea balta in 4.25 branch]