Fixed crash when using a loop with more than 32 iterations in custom HLSL.
The default threshold for loop unrolling was 32, so the compiler tried to make a dynamic loop when we had more iterations. The vector VM doesn't support that, so the codegen was never designed to understand dynamic loops, leading to an assertion failure. Increased the threshold to a large value, so we never hit this problem in practice.

#rnx
#jira none
#rb none

#ROBOMERGE-OWNER: ryan.gerleve
#ROBOMERGE-AUTHOR: mihnea.balta
#ROBOMERGE-SOURCE: CL 12422302 in //UE4/Release-4.25/... via CL 12422324 via CL 12426823 via CL 12426986
#ROBOMERGE-BOT: NETWORKING (Main -> Dev-Networking) (v673-12478461)

[CL 12529198 by mihnea balta in Dev-Networking branch]