- Fix for case where people try to take an automation screenshot during BeginPlay.
Now perform the pre-screenshot flush of loading on the first tick of the latent action and not in its constructor. This addresses cases where BP's attempt to take a screenshot in their BeginPlay event while the world is being created.

- Fix for crash that can happen when a window needs resized after a screenshot is captured.

Don't delete (which leads to restoring the viewport size) on response to the screenshot delegates. Instead queue it by one frame


#jira UE-91269
#rb na
#lockdown Cristina.Riveron

#ROBOMERGE-SOURCE: CL 12788643 in //UE4/Release-4.25/... via CL 12788645
#ROBOMERGE-BOT: RELEASE (Release-4.25Plus -> Release-Engine-Staging) (v681-12776863)

[CL 12788646 by andrew grant in Release-Engine-Staging branch]