#jira none
Fix for issue where the Oscilloscope NDI not working in PIE when Create New Audio Device For PIE is enabled in the Editor Settings. Issue was that the fix for the Oscilloscope only using half the buffer wasn't properly set up for FNiagaraDataInterfaceProxyOscilloscope::OnNewDeviceCreated the same way it is in FNiagaraDataInterfaceProxyOscilloscope::RegisterToAllAudioDevices(). To prevent future errors like this, I'm going to consolidate these to one function in Main.

[FYI] dan.reynolds, phil.popp
#rb none
#lockdown nick.whiting

#ROBOMERGE-SOURCE: CL 12908966 in //UE4/Release-4.25/... via CL 12908968 via CL 12908969
#ROBOMERGE-BOT: RELEASE (Release-Engine-Staging -> Main) (v682-12900288)

[CL 12908970 by ethan geller in Main branch]