Edigrated CL 13014082
Fixed crash in HISM. OnPostLoadPerInstanceData was properly detecting and calling a force BuildTree, but the build was failing because the static mesh of the HISM was null. For an unknown reason, data was corrupted but FoliageType's static mesh was still valid. We now detect this rare case and fix-up the component in AInstancedFoliageActor::PostLoad (editor-only).

#rb patrick.enfedaque
#jira UE-92357

[CL 13014271 by Richard Malo in Main branch]