Unified handling of element count in RHICreateShaderResourceView.
Each RHI was doing its own validation of the requested SRV size, which resulted in inconsistent behavior: some were clamping the count to make sure it fits in the underlying buffer, some were checking the offset, some weren't doing anything. In particular, Vulkan and Metal didn't know that a size of UINT32_MAX meant map the entire buffer, which led to problems, as the high level code recently started using this feature. Moved all this into a common place to get consistent behavior everywhere (namely, the element count is clamped to make sure the view doesn't extend past the end of the buffer; this was what D3D11 and 12 were doing already).

#rnx
#rb Jonas.Meyer

#ROBOMERGE-SOURCE: CL 13046654 via CL 13046656 via CL 13046658
#ROBOMERGE-BOT: RELEASE (Release-Engine-Staging -> Main) (v686-13045012)

[CL 13048473 by mihnea balta in Main branch]