[Pull from Github] FileSystemBlockSize padding would erroneously consider a file that ended at the end of a block (e.g. a 4095 byte file at offset 1 in a 4kB block) as crossing the block boundary and needing padding - by a padding amount equal to the file size.
RequiredPatchPadding had the opposite issue - a file that started on byte 0 of a block would be treated as crossing the boundary that was before the file (a 2 byte file at offset 0 was claimed to be crossing the boundary somehow, despite being literally the first two bytes in the pak file) - and a file that ended on byte 0 of the next block would be treated as not crossing the boundary. e.g. a 2 byte file at offset 4095 in the block (written to indexes 4095-4096 inclusive) would be treated as not crossing a boundary - despite having one byte in each block!

PR #6907: Fix some bad logic for calculating when padding is needed in Pak files (Contributed by LizardThief)

#rb Hongyi.Yu

#ROBOMERGE-OWNER: ryan.gerleve
#ROBOMERGE-AUTHOR: hongyi.yu
#ROBOMERGE-SOURCE: CL 13252159 in //UE4/Main/...
#ROBOMERGE-BOT: NETWORKING (Main -> Dev-Networking) (v696-13331060)

[CL 13377383 by hongyi yu in Dev-Networking branch]