Sequencer: Sequence hierarchy navigation
There's now a new picker to show the sequence hierarchy which allows you to pick any subsequence to dive into directly. There's also forward and backwards navigation buttons. And subsequences are now always shown even if there's a current filter so that it's easy to move between subsequences with the same filter.

#rb matt.hoffman

#ROBOMERGE-SOURCE: CL 13424320 via CL 13424332 via CL 13424346
#ROBOMERGE-BOT: RELEASE (Release-Engine-Staging -> Main) (v696-13331060)

[CL 13424358 by max chen in Main branch]