Ensure AssetRegistry OnPathAdded and OnPathRemoved events are always fired hierarchically
Previously, if you added "/A/B/C" to the asset registry, you would only receive a notification that "/A/B/C" was added, even if "/A" and "/A/B" didn't previously exist. Conversely, if removing "/A" you would only be notified that "/A" was removed, and not also "/A/B/C" and "/A/B".

This change ensures that all paths added are notified parent -> child at each level, and that each path removed is notified child -> parent. Additionally mount roots are now added immediately rather than waiting for content to exist within them, which ensures that they always exist in empty projects.

#rb Francis.Hurteau

#ROBOMERGE-SOURCE: CL 13545754 via CL 13545762 via CL 13545774
#ROBOMERGE-BOT: RELEASE (Release-Engine-Staging -> Main) (v700-13537989)

[CL 13545780 by jamie dale in Main branch]