Create large fixed-sized ray tracing view descriptor heaps instead of creating worst-case heaps for the current frame. Initial heap size is set by r.D3D12.RayTracingViewDescriptorHeapSize (default = 250k).
This avoids the pathological performance issue when new meshes are progressively added to a large ray tracing scene. The previous allocation scheme resulted in ever-larger heaps being created as the scene grew, which caused significant CPU spikes (200-300 ms). New scheme reuses the same heaps that are pre-sized to be large enough. If overflow is detected, the heap size will be doubled for the next frame, but the current frame will have some number of placeholder shaders bound to the ray tracing scene and will not be rendered correctly as the result.

#rb Juan.Canada
#ushell-cherrypick of 13539503 by Yuriy.ODonnell

#ROBOMERGE-OWNER: jeanmichel.dignard
#ROBOMERGE-AUTHOR: yuriy.odonnell
#ROBOMERGE-SOURCE: CL 13608503 via CL 13608504 via CL 13608507 via CL 13608511
#ROBOMERGE-BOT: TOOLS (Main -> Dev-Tools-Staging) (v703-13590297)

[CL 13610853 by yuriy odonnell in Dev-Tools-Staging branch]