Track hovered and focused viewport clients in the Editor Mode Tools
The InteractiveGizmo framework needs to keep track of both the focused and the currently active (could be hovered) viewport clients.  In the level editor this was done through a global variable which tracks the "current" viewport client.  This never really worked though in multi viewport situations. Now that we want the gizmos to work in asset editors as well, the global variable won't work at all.  The EditorModeTools are in the right place to keep track of this information.  This change is a first stab at (naively) tracking thiese VPClients.  We can refine it in the future, but it works decently for now.

#rnx
#jira:none
#rb ryan.schmidt
#rb lauren.barnes

#ROBOMERGE-SOURCE: CL 13651262 via CL 13651267 via CL 13651268
#ROBOMERGE-BOT: RELEASE (Release-Engine-Staging -> Main) (v707-13641620)

[CL 13651269 by michael daum in Main branch]