Augment FTCHARToUTF8_Convert to return the number of bytes written, and to explicitly indicate if the full string could not be written. Opted for a surgical strike rather than a larger refactor to return this info.
Done to support forthcoming optimizations to IAnalyticsProviderET. By avoiding precomputing the UTF8 string length (which is complicated due to surrogate pair logic), the Analytics flush code could more than double the speed of a naive UTF8 conversion. Instead, we optimistically size the output buffer and keep expanding it until it fits, falling back to computing the actual length if a few attempts fail. In practice, this never happens for analytics because almost nothing non-ASCII is sent.

Also exposed the generic version of ::Convert (which it was already using under the hood) to allow calling code to supply their own output iterator. In theory it allows for precise buffer sizing without having to guess the proper size, while still only scanning the string once. In practice it was 50% slower than optimistically guessing the size because of the overhead of Add()'ing one element at a time to the OutputBuffer that was required to safely use an output iterator.

Including the last folks involved in UTF8 changes.

#rb none
#tests thorough testing using accompanying IAnalyticsProviderET changes.
[FYI] ian.fox, jamie.dale, justin.sargent, steve.robb

#ROBOMERGE-SOURCE: CL 13695467 via CL 13695652 via CL 13695689
#ROBOMERGE-BOT: RELEASE (Release-Engine-Staging -> Main) (v707-13641620)

[CL 13695717 by wes hunt in Main branch]