Sequencer: Evaluation tree population logic changes and cleanup.
In the pre-ECS system it was defaulting to a high-pass-per-row algorithm that nonetheless added any section that had a valid blend type, probably based on the idea that if something has a blend type, it can probably handle being blended with other sections in the same time segment.

ECS was now defaulting to a logic where everything was added to the evaluation tree. Because most things in Sequencer support blending, it was very close to the same result, but not quite the same. This change ports the old logic to the new API.

The change also:
- moves all the default population rules in a separate file and utility class
- makes the sequencer track "PopulateEvaluationTree" method not take a list of sections, since the track already knows of its own sections (this removes potential confusion for track implementers)

#rb andrew.rodham


#ROBOMERGE-SOURCE: CL 13792772 via CL 13792773
#ROBOMERGE-BOT: (v717-13788717)

[CL 13792776 by ludovic chabant in Release-Engine-Staging branch]