Addendum to fixing, "Assertion failed: !Obj->HasAnyFlags(RF_NeedLoad|RF_NeedPostLoad|RF_ClassDefaultObject) || bIsOwnedByCDOOrArchetype".
Fixing a crash in the designer, due to the way we create the DragDrop Previews they were *all* being outered to the WidgetBlueprint WidgetTree, instead of either the template tree or the preview tree.  This lead to a situation for some people, oddly not everyone always - where we'd end up trying to initiailize slate widgets on an un-initialized UUserWidget, because that user widget was outered to an archetype WidgetTree, it had no business being initialized, that's only something we do for full fleged widget trees and userwidgets we're expanding.

Needless to say there was more fallout, it appears we've been getting by with this hack in some places, e.g. UCanvasPanelSlot::DragDropPreviewByDesigner was just straight up calling TakeWidget on what could / should be an uninitiailized UserWidget.

For now I've disabled that feature so that we don't crash, but need to see if this is perhaps just allowing (initializing userwidgets inside an archetype tree) my gut says it's a bad idea, that was never intended to work, the fact that some systems were doing it but we lacked enforcement is concerning though - though probably only extends as far as these drag/drop issues in the designer.

[at]Vincent.Gauthier, [at]Patrick.Boutot
#tests passed local testing

#ROBOMERGE-OWNER: ryan.gerleve
#ROBOMERGE-AUTHOR: nick.darnell
#ROBOMERGE-SOURCE: CL 13848681 via CL 13848688 via CL 13848693 via CL 13848696 via CL 13848704
#ROBOMERGE-BOT: NETWORKING (Main -> Dev-Networking) (v717-13788717)

[CL 13849795 by nick darnell in Dev-Networking branch]