#jira UE-93034
Editor layout file crashing editor --> FLayoutSaveRestore::LoadFromConfig() was not considering the different cases of "EOutputCanBeNullptr" (in particular for MainFrameModule.cpp), in which a layout could be wrong if no open tabs were valid or present. This particular layout did not have an opened tab. Added a 3rd argument to LoadFromConfig(), "bPrimaryAreaMustHaveOpenedTabsToBeValid", to make it match with the right "EOutputCanBeNullptr" argument of FGlobalTabmanager::Get()->RestoreFrom().

Also, if an unknown error occurs while reading a layout file, that layout was being removed. Now, the layout is simply renamed to "EditorLayout_faulty.ini", so it can be debugged later.

#rb none

[CL 13867266 by Gines Hidalgo in Main branch]