#jira UE-85371
If a loaded Layout does not contain the requested field but contains an older one, we remove the old one and warn the user (UE_LOG Warning + UI warning).

Issue: Changes in the version of the components of the layout INI result in a layout reset. E.g., some changes were done in "LevelEditor_Layout", which was upgraded from LevelEditor_Layout_v1.1 to LevelEditor_Layout_v1.2. Every time a component from the Layout ini file is updated, it becomes not back-compatible and that part of the layout is reset to the default layout, resulting sometimes in ugly-looking UIs.

Technical details:
Updated FLayoutSaveRestore::LoadFromConfig() with 2 new arguments:  "const bool bRemoveOlderLayoutVersions, TArray<FString>& OutRemovedOlderLayoutVersions". If bRemoveOlderLayoutVersions is true, it checks for older versions when the Key is not found, removes them, and return their names on OutRemovedOlderLayoutVersions, which is used later for the UI warning for the user.

#rb none

[CL 14062349 by Gines Hidalgo in Dev-Tools-Staging branch]