Chaos - fix crash in SyncBodies when Solver is disabled.
When a solver is disabled, we still enqueue commands to create/destroy objects in the physics scene, but the commands are not executed until the solver is enabled. However, we were still attempting to pull data from the solver, even when it was disabled.

NOTE: create/destroy particle commands will accumulate while solver is disabled. This will be addressed separately.

#rb ori.cohen

#ROBOMERGE-OWNER: ryan.gerleve
#ROBOMERGE-AUTHOR: chris.caulfield
#ROBOMERGE-SOURCE: CL 14028837 via CL 14028840 via CL 14028843 via CL 14028850
#ROBOMERGE-BOT: NETWORKING (Main -> Dev-Networking) (v718-13788717)

[CL 14072874 by chris caulfield in Dev-Networking branch]