Initial "typed element" support
This is a framework which provides a way to access an "element" via a series of interfaces, where each type of element may or may not choose to implement a particular interface.

An element type may either be "typed" or "untyped" depending on whether it requires extra payload information to store itself, or whether it can be represented purely by the element ID. The allocation and management of payload instances is handled internally as required, and the payload itself can be any C++ struct that uses the UE_DECLARE_TYPED_ELEMENT_DATA_RTTI and UE_DEFINE_TYPED_ELEMENT_DATA_RTTI macros to add minimal element payload RTTI data.

Client code will most commonly make use of FTypedElementHandle (a combination of an element ID and its internal data pointer) and TTypedElement (a combination of an element handle and an interface instance that is associated with it). FTypedElementList is useful for storing a set of elements (eg, a selection set), and FTypedElementId may be useful in limited circumstances where you need the most minimal representation of an element (current 32-bits).

Implementation code will need to make use of UTypedElementInterface to create element interface types, as well as UTypedElementRegistry to register, create, and destroy instances of an element. An element created via UTypedElementRegistry is returned as a TTypedElementOwner, which provides mutable access to the internal payload data (if any), and is the only time outside of implementing an interface for a specific element type that you should make any assumptions about the type of data held within an element instance.

TypedElementFrameworkTests.h/.cpp provides a basic example of usage. Note: While UTypedElementRegistry isn't intrinsically a singleton, there is a global instance available via UTypedElementRegistry::GetInstance().

#rb Chris.Gagnon

[CL 14124587 by Jamie Dale in ue5-main branch]