Converted UGS into a Net Core application
* Mostly removed unused references which didnt exist in net core.
* Replaced usages of JsonScriptSerializer (System.Web) with the new System.Text.Json (as the old one does not exist in net core).
* Updated installer to install Launcher and publish it as a self contained application (Makes msi 64 MB instead of 600kb but does not require anything to be installed now).

Worked around what looks to be a bug in WinForms for Net Core, in Workspace Control the event state was selected when drawing the item even though the item was not selected (and the state said it isnt selected). Reading the state from the object instead of looking at the event.

Deleted ProjectWindow form as this was never used and code did not compile (net core projects builds all files without them having to be included)

Deleted metadata server as this has been replaced by Horde. Also removed Post Build Status tool for the same reason.

#rb ben.marsh

[CL 14153139 by Joakim Lindqvist in ue5-main branch]