Update FMacPlatformProcess::GetDllHandle() to accommodate macOS 11 Big Sur
On macOS 11, the system-provided dynamic libraries do not exist on the filesystem, only in a built-in dynamic linker cache.  Any code that attempts to check for dynamic library presence by looking for a file at a path or enumerating a directory will fail. Instead, we must check for library presence by attempting to dlopen() the path, which will correctly check for the library in the cache.

#rb michael.trepka andrew.grant
#jira UE-97265
#rnx


#ROBOMERGE-SOURCE: CL 14179008 via CL 14179014 via CL 14179022 via CL 14179034
#ROBOMERGE-BOT: (v720-14121032)

[CL 14179042 by will damon in Release-Engine-Staging branch]