Only close, add & reopen the default context command list to execute when it's not closed
This can currently happen when the RenderThread takes 'ownership' of the RHIThread via the FScopedRHIThreadStaller and then calls FlushCommands. This could happen while RHIThread is waiting on multiple parallel command lists to finish and then the default command list could have already been closed and added to the pending list to execute. It will be reopened when it's done waiting for the last parallel command list in a batch.
This ideally needs to be fixed on a higher level as well but this change is safe in any case, and fixes a potential GPU crash.

#rb Emil.Persson

#ROBOMERGE-SOURCE: CL 14186765 via CL 14186768 via CL 14186769 via CL 14186771 via CL 14186772
#ROBOMERGE-BOT: RELEASE (Release-Engine-Staging -> Main) (v720-14121032)

[CL 14187084 by kenzo terelst in Main branch]