Jira: Opening the Project Settings for Fortnite causes a noticeable hitch in the editor
Remove CreateIfNotFound parameter to FConfigCacheIni::Find.

This was creating thousands of fstat calls peeking for the existence of Saved/Config/WindowsEditor/GameplayTagsEditorState.ini when opening the Project Settings dialog. This we make frequent calls (in a recursive manner) into the config system, but no file or config is created (where false is passed), and we keep statting it thousands of times

Callers that pass false for this fail to land on any config properties because they don't exist to begin with.

Callers that pass true for this still basically get what they want. Since we did not stub the file out upon Find(), there is no disk state difference between Find() and config system shutdown (where dirty config files are saved to disk). So saving is still manual, and there is no implication from
CreateIfNotFound communicated to the FConfigFile about whether it is meant to be saved or not. So by and large this semantic is meaningless except with the existence of the new config file of the config file map (which creates the optimization of skipping the subsequent stats).

SSettingsEditor::ReloadCategories before: 2290ms
SSettingsEditor::ReloadCategories after: 1669ms

Tested with a BuildCookRun of ShooterGame, and running with it in Editor for a few days.

#jira UE-95853

[CL 14258235 by geoff evans in ue5-main branch]