Mesh-to-mesh skinning: support multiple influence triangles.
Before this change, a high-res render mesh vertex is bound to a single low-res simulation mesh triangle by barycentric weighting. This change allows the use of multiple low-res/sim triangles to influence a high-res mesh vertex. Similar to Houdini's Cloth Capture and Cloth Deform node, we use a polynomial kernel to weight the influence of the multiple triangles.

Limitations (should be fairly easy to change these):
- hard coded limit of 5 influences per vertex
- only one hard-coded kernel available

This change also appears to address an issue with skinned normals not working correctly (UE-76156).

#rb kriss.gossart david.hill benn.gallagher alexis.matte
#jira UETOOL-2288 UETOOL-2321 UE-76156

#ushell-cherrypick of 13776557 by tyson.brochu

#ROBOMERGE-OWNER: jeanmichel.dignard
#ROBOMERGE-AUTHOR: tyson.brochu
#ROBOMERGE-SOURCE: CL 14310483 in //UE4/Release-4.26/... via CL 14313476 via CL 14313676
#ROBOMERGE-BOT: TOOLS (Main -> Dev-Tools-Staging) (v739-14336328)

[CL 14348488 by tyson brochu in Dev-Tools-Staging branch]