Merging 14269469, 14299800 Fix range selection in Persona treeview when initial selection is made from the viewport, and the second from the treeview.
Note: This introduces a breaking API change for classes inheriting IPersonaPreviewScene or ISkeletonTree that override SetSelectedBone, and for user code that binds to their SelectionChanged event. These now take an extra argument to route the cause of the change (user interaction or direct code) through to the UI elements.

Previously when selection a bone/body in the viewport it would change the active selection in the treeview but would not reset the selection range start. If you then shift-selected in the tree view it would select a range starting from whatever was last select in the treeview, regardless of what was selected in the viewport. This was caused by not passing the selection info through the selection callstack, and the default selection behaviour assuming a programmatic selection (rather than a user selection) which does not reset the range start.

#rb none
#jira none

#ROBOMERGE-OWNER: jeanmichel.dignard
#ROBOMERGE-AUTHOR: chris.caulfield
#ROBOMERGE-SOURCE: CL 14319499 in //UE4/Release-4.26/... via CL 14319637 via CL 14319645
#ROBOMERGE-BOT: TOOLS (Main -> Dev-Tools-Staging) (v739-14336328)

[CL 14349932 by chris caulfield in Dev-Tools-Staging branch]