SceneViewExtensions: Added a Context that a Scene Extension can use to determine if it should be active or not.
Currently the main way that Scene View Extensions decide whether to be active in a viewport for the current frame is a pointer to FViewport.

But this is sometimes not sufficient information. For example, you may want a Scene View Extension to be active on a particular Scene Capture Component and not others. But they do not have an FViewport, so they pass null.

Or you may want a scene extension on a particular view that shares the same Viewport.

So this CL provides a way to have an arbitrary Context that corresponds to the rendering entity, and a mechanism to provide the SVE lambdas that can look for the custom context and parse it. It can then decide if it should be active for the current frame and context, or simply not have an opinion (and would fall back to usual behavior).

This does not change the normal behavior, and only adds an option for renderers to provide a context.

#rb Jonathan.Bard, Kevin.Ortegren
#jira UE-99465

#ushell-cherrypick of 14330850 by Alejandro.Arango

#ROBOMERGE-SOURCE: CL 14362946 in //UE4/Release-4.26/... via CL 14362949
#ROBOMERGE-BOT: RELEASE (Release-Engine-Staging -> Main) (v741-14365223)

[CL 14367314 by alejandro arango in Main branch]