Details view reskin phase 1.
The details view now displays gridlines between rows, and now has 4 columns - edit condition, name, value, and reset/extensions.

Edit condition widgets are now created in SDetailSingleItemRow and are placed in their own fixed column on the left of the details view. FEditConditionParser is now in FPropertyNode rather FPropertyEditor, which allows access to it more broadly - notably in SDetailSingleItemRow.

The Reset to Default SWidget can no longer be replaced - the logic for it being visible and a handler for when it is pressed is still customizable. SPropertyEditorAsset no longer has a special slot for the reset to default widget.

All extension widgets are now placed in a resizable column on the right of the panel, including the Keyframe button.

FDetailColumnSizeData is now publicly exposed by IDetailsView, and is no longer passed around as much.

#rb matt.kuhlenschmidt

[CL 14375859 by sebastian nordgren in ue5-main branch]