Cameras: Move camera shake implementation in a "shake pattern" class, leaving the base class as a container.
This is to make it easier for artists to create placeholder shakes. The new shake would use, say, a noise or sinewave pattern, and later a designer can replace that with a mocap'ed animation using a Sequencer shake pattern, without having to change all the places that used the placeholder.

The base camera shake is not creatable because it doesn't have good default. A subclass that defaults to the perlin noise pattern will be available in the upcoming GameplayCameras plugin as the default camera shake class. The older Matinee camera shake also appears as an available base class for backwards compatibility reasons while not everything is in place.

#rb max.chen

#ROBOMERGE-SOURCE: CL 14393362 via CL 14393371 via CL 14396703
#ROBOMERGE-BOT: RELEASE (Release-Engine-Staging -> Main) (v742-14376270)

[CL 14397514 by ludovic chabant in Main branch]