Sequencer: Crossing a preroll boundary now reimports entities in-place rather than completely destroying them
This ensures that the state is maintained for such entities across prerolls (ie entities with a spawnable binding component on them will not cause the spawnable to be destroyed and recreated again). In the case of spawnables this actually manifested itself by first (re)spawning the same spawnable (causing it to simply early return since it was already spawned), then destroy the preroll spawnable, causing the whole object to be destroyed.

The one edge case that this could open up is where a preroll entity hass significantly different components from its normal entity, and there are systems that track the link/unlink lifetime for such components. Under this circumstance, the unlink for the entity will not be seen until the whole entity goes away, not just the preroll. This is a very nuanced edge-case however, and can be solved by giving the entities unique IDs in the evaluation field. Ultimately we could handle these cases better by supporting component lifetime events in the sequencer ecs.

#jira UE-100564
#rb Max.Chen, Ludovic.Chabant, Matt.Hoffman

#ROBOMERGE-SOURCE: CL 14437557 in //UE4/Release-4.26/... via CL 14437562
#ROBOMERGE-BOT: RELEASE (Release-Engine-Staging -> Main) (v745-14436861)

[CL 14437566 by andrew rodham in Main branch]