MoviePipeline: This adds a stencil-based layering system to the deferred render passes. This works by marking each object on the layer to write into the stencil mask and then using a post processing material to write pixels that fall outside the stencil mask as black and non-transparent (which then gets inverted later).
This supports post processing (TAA, DoF, MotionBlur) and the layers can be addtively stacked together in a compositing software to re-create the final image. Each layer contains only the objects on that layer and are masked by foreground objects which means the order they are combined does not matter, but it also means that a 'background' layer will have a black cutout where foreground objects occlude it. An alpha channel will be written which can be used as a mask to re-extract the pixels that were actually touched by this layer which will enable doing grading adjustments like gain.

Translucency is partially supported - a translucent object will show the combined translucent + opaque pixels if you have both translucent + opaque objects in the same layer. Translucent pixels that overlap another layer will be rendered against black, but combining them with addition (A+B+C) is not the same math as blending them so the look will change.

Only works with materials that can write to the stencil layer.

The stencil based approach was chosen because of limitations of deferred rendering. With deferred rendering, things that are not rendered are not present in many of the effects - shadow casters won't cast shadows, screen space (and ray traced) reflections won't reflect the non-rendered objects, etc. By rendering the entire scene and then masking out the pixels you don't want all of these objects will contribute to the final render. The tradeoff to this is cost (rendering the entire scene), and foreground objects still obscuring background objects.

#jira UESEQ-715, UESEQ-645
#rb Rod.Bogart

#ROBOMERGE-OWNER: jeanmichel.dignard
#ROBOMERGE-AUTHOR: matt.hoffman
#ROBOMERGE-SOURCE: CL 14448509 in //UE4/Release-4.26/... via CL 14448532 via CL 14448561
#ROBOMERGE-BOT: TOOLS (Main -> Dev-Tools-Staging) (v748-14444958)

[CL 14456103 by matt hoffman in Dev-Tools-Staging branch]